+++ app/controllers/registrations_controller.rb
@@ -7,17 +7,17 @@ class RegistrationsController < ApplicationController
   def create
     @user = User.new(user_params)
 
-    if @user.save
+    if @user.save(context: :signup)
       session[:user_id] = @user.id
       redirect_to root_path, notice: "Welcome, #{@user.name}!"
     else
-      render :new, status: :unprocessable_entity
+      render :new, status: :unprocessable_content
     end
   end
 
   private
 
   def user_params
-    params.require(:user).permit(:username, :email, :display_name, :password, :password_confirmation)
+    params.require(:user).permit(:first_name, :last_name, :username, :email, :password, :password_confirmation)
   end
 end