+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2026_01_01_000012) do
+ActiveRecord::Schema[7.1].define(version: 2026_01_01_000013) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -89,10 +89,11 @@ ActiveRecord::Schema[7.1].define(version: 2026_01_01_000012) do
     t.string "username", null: false
     t.string "email"
     t.string "password_digest", null: false
-    t.string "display_name"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.boolean "admin", default: false, null: false
+    t.string "first_name"
+    t.string "last_name"
     t.index ["admin"], name: "index_users_on_admin"
     t.index ["email"], name: "index_users_on_email", unique: true, where: "(email IS NOT NULL)"
     t.index ["username"], name: "index_users_on_username", unique: true